Fixed cache inconsistancy issue on writes.
When one or more documents are updated, the cache is updated with the
new documents. The documents which are accessed during the save process
are also added to the cache if they are not already present. The order
of adding these two sets was wrong and that was causing the trouble.