Fixed error in moving types.
When types are being moved, infogami collects all types from the system
and passes it to save_many as a list. When two items with same key
appear in save_many list, infobase used to pick the last one. It doesn't
seem to be happening now. Converted the list of items into a dict and
taking its values to get the expected behavior.